kill the taxonomy trait
